// src/modules/vendors/onboarding/enums/vendor-onboarding-step.enum.ts
export enum VendorOnboardingStep {
  VENDOR_KYC_VERIFIED = 'VENDOR_KYC_VERIFIED',
  SETTLEMENT_ACCOUNT = 'SETTLEMENT_ACCOUNT',
  BUSINESS_CREATED = 'BUSINESS_CREATED',
  BUSINESS_KYC_VERIFIED = 'BUSINESS_KYC_VERIFIED',
  MENU_CREATED = 'MENU_CREATED',
  AVAILABILITY_SET = 'AVAILABILITY_SET',
  READY_TO_GO_LIVE = 'READY_TO_GO_LIVE',
}